pulling_data/requirements.txt
flask
pandas
numpy
tweepy
praw
XlsxWriter
requests
python-dotenv

